Possible solution : Make template always visible
Note that we're not using EXPORTTEST_API, which would make it visible only when compiling libExportTest. We're using EXPORTTEST_EXPORT, which overrides the visibility to always be default, no matter what we're compiling. We can then use the fact that the implementation is hidden to control which cpp files we generate the symbols into.