Expose intrusive_ptr symbol visibility issue
This models another problem that has cropped up when managing symbol visibility for Cortex on Linux. The RefCountedPtr is a typedef for boost::intrusive_ptr<RefCounted>, and the boost header doesn't specify any visibility settings, therefore it inherits the hidden visibility set by the compiler. We then try to instantiate `TypedData<map<int, RefCountedPtr>>` but the compiler gives it hidden visibility too, even though we told it not to, because the RefCountedPtr part of the type is hidden.